@@ -16,7 +16,8 @@
     <MinVerBuildMetadata>build.$(BUILD_NUMBER)</MinVerBuildMetadata>
     <PackageTags>aspnetcore;proxy;http</PackageTags>
     <NoWarn>$(NoWarn);NU5105;1701;1702;1705;1591</NoWarn>
-    <LangVersion>7.3</LangVersion>
+    <LangVersion>8</LangVersion>
+    <EnableNullable>true</EnableNullable>
     <DebugType>full</DebugType>
     <PackageIconUrl>https://raw.githubusercontent.com/damianh/ProxyKit/master/logo.png</PackageIconUrl>
     <PackageReleaseNotes>See https://github.com/damianh/ProxyKit/releases for release notes.</PackageReleaseNotes>
