@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net.Http;
+
 using Microsoft.Extensions.DependencyInjection;
 
 namespace ProxyKit
@@ -13,7 +14,7 @@ namespace ProxyKit
             Action<IHttpClientBuilder> configureHttpClientBuilder = null,
             Action<ProxyOptions> configureOptions = null)
         {
-            if (services == null)
+            if (services is null)
             {
                 throw new ArgumentNullException(nameof(services));
             }
@@ -28,7 +29,7 @@ namespace ProxyKit
 
             configureHttpClientBuilder?.Invoke(httpClientBuilder);
 
-            configureOptions = configureOptions ?? (_ => { });
+            configureOptions ??= (_ => { });
             services
                 .Configure(configureOptions)
                 .AddOptions<ProxyOptions>();
