@@ -1,15 +1,15 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+
 using Microsoft.AspNetCore.Http;
 
 namespace ProxyKit
 {
-    public class ProxyMiddleware<TProxyHandler> where TProxyHandler:IProxyHandler
+    public sealed class ProxyMiddleware<TProxyHandler> where TProxyHandler : IProxyHandler
     {
         private readonly TProxyHandler _handler;
         private const int StreamCopyBufferSize = 81920;
@@ -55,7 +55,7 @@ namespace ProxyKit
                     await responseStream.CopyToAsync(response.Body, StreamCopyBufferSize, context.RequestAborted).ConfigureAwait(false);
                     if (responseStream.CanWrite)
                     {
-                        await responseStream.FlushAsync(context.RequestAborted).ConfigureAwait(false);    
+                        await responseStream.FlushAsync(context.RequestAborted).ConfigureAwait(false);
                     }
                 }
             }
